Move predicates types to separate module
