+++ src/client/kotlin/liltojustice/trueadaptivemusic/client/gui/widget/PredicateViewWidget.kt
@@ -1,10 +1,10 @@
 package liltojustice.trueadaptivemusic.client.gui.widget
 
 import liltojustice.trueadaptivemusic.client.MusicPack
-import liltojustice.trueadaptivemusic.client.predicate.MusicPredicate
 import liltojustice.trueadaptivemusic.client.predicate.MusicPredicateTree
-import liltojustice.trueadaptivemusic.client.predicate.RootPredicate
+import liltojustice.trueadaptivemusic.client.predicate.types.RootPredicate
 import liltojustice.trueadaptivemusic.client.identifier.TypedIdentifier
+import liltojustice.trueadaptivemusic.client.predicate.types.MusicPredicate
 import liltojustice.trueadaptivemusic.client.sound.PlayableSound
 import liltojustice.trueadaptivemusic.client.sound.PlayableSoundEvent
 import net.minecraft.client.gui.DrawContext
