+++ src/client/kotlin/liltojustice/trueadaptivemusic/client/predicate/MusicPredicateTree.kt
@@ -4,6 +4,8 @@ import com.google.gson.JsonArray
 import com.google.gson.JsonObject
 import liltojustice.trueadaptivemusic.LogLevel
 import liltojustice.trueadaptivemusic.Logger
+import liltojustice.trueadaptivemusic.client.predicate.types.MusicPredicate
+import liltojustice.trueadaptivemusic.client.predicate.types.RootPredicate
 import liltojustice.trueadaptivemusic.client.sound.PlayableSound
 import liltojustice.trueadaptivemusic.client.sound.PlayableSoundEvent
 import liltojustice.trueadaptivemusic.client.sound.PlayableSoundFile
@@ -18,10 +20,6 @@ typealias NodeVisitor = (MusicPredicateTree.Node, Int) -> Unit
 class MusicPredicateTree private constructor(
     json: JsonObject? = null, private val soundLibrary: Map<String, PlayableSoundFile> = mapOf()) {
     private val root = if (json != null) Node.fromJson(json, soundLibrary) else Node.makeRoot()
-
-    fun copy(): MusicPredicateTree {
-        return MusicPredicateTree(toJson(), soundLibrary.toMap())
-    }
 
     fun toJson(): JsonObject {
         return root.toJson()
