+++ src/client/kotlin/liltojustice/trueadaptivemusic/client/predicate/types/DimensionPredicate.kt
@@ -1,4 +1,4 @@
-package liltojustice.trueadaptivemusic.client.predicate
+package liltojustice.trueadaptivemusic.client.predicate.types
 
 import com.google.gson.JsonObject
 import com.google.gson.JsonPrimitive
