+++ src/client/kotlin/liltojustice/trueadaptivemusic/client/predicate/types/TitleScreenPredicate.kt
@@ -1,4 +1,4 @@
-package liltojustice.trueadaptivemusic.client.predicate
+package liltojustice.trueadaptivemusic.client.predicate.types
 
 import com.google.gson.JsonObject
 import net.minecraft.client.MinecraftClient
